# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: cmp-security是一个客户关系管理权限平台
  version: 1.0.0
  title: 客户关系管理权限平台
basePath: /cmp-security
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/menu/info/{menuId}:
    get:
      tags:
        - 菜单管理
      summary: 获取菜单信息
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 菜单ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回菜单信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              menu:
                description: 菜单信息
                $ref: '#/definitions/SysMenuEntity'
  /sys/menu/save:
    post:
      tags:
        - 菜单管理
      summary: 添加菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/update:
    post:
      tags:
        - 菜单管理
      summary: 修改菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/sysmenuview/list:
    get:
      tags:
        - 菜单视图管理
      summary: 菜单视图列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: menuId
          description: 菜单id
          in: query
          type: integer
        - name: viewName
          description: 视图名称
          in: query
          type: string
        - name: isView
          description: 是否默认
          in: query
          type: string
        - name: orderNum
          description: 排序号
          in: query
          type: integer
        - name: createCode
          description: 创建人代码
          in: query
          type: string
        - name: createName
          description: 创建人
          in: query
          type: string
        - name: startCreateTime
          description: 开始创建时间
          in: query
          type: string
          format: date
        - name: endCreateTime
          description: 结束创建时间
          in: query
          type: string
          format: date
        - name: startCreateTime
          description: 开始创建时间
          in: query
          type: string
          format: date
        - name: endCreateTime
          description: 结束创建时间
          in: query
          type: string
          format: date
        - name: updateCode
          description: 修改人代码
          in: query
          type: string
        - name: updateName
          description: 修改人
          in: query
          type: string
        - name: startUpdateTime
          description: 开始修改时间
          in: query
          type: string
          format: date
        - name: endUpdateTime
          description: 结束修改时间
          in: query
          type: string
          format: date
      responses:
        '200':
          description: 返回菜单视图列表
          schema:
            $ref: '#/definitions/SysMenuViewEntityList'
  /sys/sysmenuview/info/{id}:
    get:
      tags:
        - 菜单视图管理
      summary: 获取菜单视图信息
      produces:
        - application/json
      parameters:
        - name: id
          description: ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回菜单视图信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysMenuViewEntity'
  /sys/sysmenuview/save:
    post:
      tags:
        - 菜单视图管理
      summary: 添加菜单视图
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单视图对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuViewEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/sysmenuview/update:
    post:
      tags:
        - 菜单视图管理
      summary: 修改菜单视图
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单视图对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuViewEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/sysmenuview/delete:
    post:
      tags:
        - 菜单视图管理
      summary: 删除菜单视图
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单视图ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/sysmenuoperation/list:
    get:
      tags:
        - 菜单操作管理
      summary: 菜单操作列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: menuId
          description: 菜单id
          in: query
          type: integer
        - name: operateCode
          description: 操作代码
          in: query
          type: string
        - name: 操作名称
          description: operationName
          in: query
          type: string
        - name: orderNum
          description: 排序号
          in: query
          type: integer
        - name: createCode
          description: 创建人代码
          in: query
          type: string
        - name: createName
          description: 创建人
          in: query
          type: string
        - name: startCreateTime
          description: 开始创建时间
          in: query
          type: string
          format: date
        - name: endCreateTime
          description: 结束创建时间
          in: query
          type: string
          format: date
        - name: startCreateTime
          description: 开始创建时间
          in: query
          type: string
          format: date
        - name: endCreateTime
          description: 结束创建时间
          in: query
          type: string
          format: date
        - name: updateCode
          description: 修改人代码
          in: query
          type: string
        - name: updateName
          description: 修改人
          in: query
          type: string
        - name: startUpdateTime
          description: 开始修改时间
          in: query
          type: string
          format: date
        - name: endUpdateTime
          description: 结束修改时间
          in: query
          type: string
          format: date
      responses:
        '200':
          description: 返回菜单操作列表
          schema:
            $ref: '#/definitions/SysMenuViewEntityList'
  /sys/sysmenuoperation/info/{id}:
    get:
      tags:
        - 菜单操作管理
      summary: 获取菜单操作信息
      produces:
        - application/json
      parameters:
        - name: id
          description: ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回菜单操作信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysMenuViewEntity'
  /sys/sysmenuoperation/save:
    post:
      tags:
        - 菜单操作管理
      summary: 添加菜单操作
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单操作对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuViewEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/sysmenuoperation/update:
    post:
      tags:
        - 菜单操作管理
      summary: 修改菜单操作
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单操作对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuViewEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/sysmenuoperation/delete:
    post:
      tags:
        - 菜单操作管理
      summary: 删除菜单操作
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单操作ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  SysMenuEntity:
    type: object
    properties:
      menuId:
        description: 菜单ID
        type: integer
        format: int64
      operateCode:
        description: 操作代码
        type: string
      defaultUrl:
        description: 默认url     0：否   1：是
        type: integer
        format: int32
      name:
        description: 菜单名称
        type: string
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      parentName:
        description: 父菜单名称
        type: string
      url:
        description: 菜单URL
        type: string
      perms:
        description: 授权标识
        type: string
      type:
        description: 类型     0：目录   1：菜单   2：按钮
        type: integer
        format: int32
      toolBar:
        description: 是否为工具菜单类型     0：否   1：是
        type: integer
        format: int32
      icon:
        description: 菜单图标
        format: int32
      orderNum:
        description: 排序
        type: integer
        format: int32
      open:
        description: 是否展开   true：展开   false：不展开
        type: boolean
        format: int32
  SysMenuEntityEdit:
    type: object
    properties:
      menuId:
        description: 菜单ID
        type: integer
        format: int64
      name:
        description: 菜单名称
        type: string
      operateCode:
        description: 操作代码
        type: string
      iconType:
        description: 图标类型
        type: integer
        format: int32
      defaultUrl:
        description: 默认url     0：否   1：是
        type: integer
        format: int32
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      url:
        description: 菜单URL
        type: string
      perms:
        description: 授权标识
        type: string
      type:
        description: 类型     0：目录   1：菜单   2：按钮
        type: integer
        format: int32
      icon:
        description: 菜单图标
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
...
  SysMenuViewEntity:
    type: object
    properties:
      id:
        description: ID
        type: integer
        format: int64
      menuId:
        description: 菜单id
        type: integer
        format: int64
      viewName:
        description: 视图名称
        type: string
      viewUrl:
        description: 视图url
        type: string
      isView:
        description: 是否默认
        type: string
      orderNum:
        description: 排序号
        type: integer
        format: int32
      createCode:
        description: 创建人代码
        type: string
      createName:
        description: 创建人
        type: string
      createTime:
        description: 创建时间
        type: string
      updateCode:
        description: 修改人代码
        type: string
      updateName:
        description: 修改人
        type: string
      updateTime:
        description: 修改时间
        type: string
  SysMenuViewEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysMenuViewEntity'
...